:recycle: [#18] -- refactor builder form zod schema building
Centralized common schema validation such as label + key, in
preparation for complexer component.type specific validations, using
zod unions and intersections.
